Refactor voucher discount to double and improve payment UI
Changed voucher discountValue from int to double throughout the codebase for better precision and flexibility. Improved payment method selection UI by removing RadioGroup in favor of direct InkWell handling, and fixed related logic in checkout and webview screens. Enhanced voucher parsing and error handling, updated rating and survey screens for better UX, and added new localized strings for rating points. Also included minor bug fixes and code cleanup.

+++ lib/objects/voucher_children/limited_amount_voucher_with_end_time.dart
@@ -59,7 +59,7 @@ class LimitedAmountVoucherWithEndTime
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    int? discountValue,
+    double? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
