+++ lib/objects/voucher_children/limited_percentage_voucher_with_end_time.dart
@@ -68,7 +68,7 @@ class LimitedPercentageVoucherWithEndTime
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    int? discountValue,
+    double? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
