+++ lib/objects/voucher_children/unlimited_percentage_voucher_with_end_time.dart
@@ -47,7 +47,7 @@ class UnlimitedPercentageVoucherWithEndTime extends Voucher
     String? voucherID,
     String? voucherName,
     DateTime? startTime,
-    int? discountValue,
+    double? discountValue,
     int? minimumPurchase,
     int? maxUsagePerPerson,
     DistributionType? distributionType,
